requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
macholib==1.16.3
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pydub==0.25.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
Werkzeug==3.1.3
